requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-extensions==3.2.1
django-jazzmin==2.6.0
django-ninja==0.21.0
django-rest-framework==0.1.0
django-silk==5.0.3
djangorestframework==3.14.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
gprof2dot==2022.7.29
idna==3.4
MarkupSafe==2.1.2
Naked==0.1.32
phonenumbers==8.13.8
Pillow==9.4.0
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.7
python-jose==3.3.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sourcery==1.1.0
sqlparse==0.4.3
status==0.2.5
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.2
urllib3==1.26.15
Werkzeug==2.2.3
